pull out transform functions for integrated corpus
